More circular import fixes
