+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include skyplane/data/aws_transfer_costs.csv
+include skyplane/data/throughput.csv
