+++ pyproject.toml
@@ -9,7 +9,7 @@ homepage = "https://skyplane.org/"
 repository = "https://github.com/skyplane-project/skyplane"
 documentation = "https://skyplane.org/"
 readme = "README.md"
-include = ["profiles/*"]
+include-package-data = true
 
 [tool.poetry.dependencies]
 python = ">=3.7.1,<3.12"
