+++ skyplane/api/impl/tracker.py
@@ -3,20 +3,23 @@ import json
 import time
 from datetime import datetime
 from threading import Thread
-from typing import Dict, List, Optional, Set
+from typing import TYPE_CHECKING, Dict, List, Optional, Set
 
 import urllib3
 
 from skyplane import exceptions
 from skyplane.api.transfer_config import TransferConfig
-from skyplane.api.impl.transfer_job import TransferJob
 from skyplane.chunk import ChunkRequest, ChunkState
 from skyplane.utils import logger, imports
 from skyplane.utils.fn import do_parallel
 
 
+if TYPE_CHECKING:
+    from skyplane.api.impl.transfer_job import TransferJob
+
+
 class TransferProgressTracker(Thread):
-    def __init__(self, dataplane, jobs: List[TransferJob], transfer_config: TransferConfig):
+    def __init__(self, dataplane, jobs: List["TransferJob"], transfer_config: TransferConfig):
         super().__init__()
         self.dataplane = dataplane
         self.jobs = {job.uuid: job for job in jobs}
