+++ skyplane/cli/cli_azure.py
@@ -13,7 +13,7 @@ from skyplane.compute.azure.azure_auth import AzureAuthentication
 from skyplane.compute.azure.azure_cloud_provider import AzureCloudProvider
 from skyplane.utils.fn import do_parallel
 from skyplane.utils import logger
-from skyplane.cli import cloud_config
+from skyplane.config_paths import cloud_config
 from rich import print as rprint
 
 from skyplane.compute.azure.azure_auth import AzureAuthentication
