+++ skyplane/cli/cli_config.py
@@ -10,7 +10,7 @@ Available keys:
 
 import typer
 
-from skyplane.cli import config_path, cloud_config
+from skyplane.config_paths import config_path, cloud_config
 from skyplane.cli.common import console
 from skyplane.cli.usage.client import UsageClient
