+++ skyplane/cli/cli_impl/init.py
@@ -10,11 +10,12 @@ import typer
 from rich.progress import Progress, SpinnerColumn, TextColumn
 import questionary
 
-from skyplane.compute.aws.aws_auth import AWSAuthentication, aws_config_path
+from skyplane.compute.aws.aws_auth import AWSAuthentication
 from skyplane.compute.azure.azure_auth import AzureAuthentication
 from skyplane.compute.azure.azure_server import AzureServer
-from skyplane.compute.gcp.gcp_auth import GCPAuthentication, gcp_config_path
+from skyplane.compute.gcp.gcp_auth import GCPAuthentication
 from skyplane.config import SkyplaneConfig
+from skyplane.config_paths import aws_config_path, gcp_config_path
 
 
 def load_aws_config(config: SkyplaneConfig, non_interactive: bool = False) -> SkyplaneConfig:
